--- packages/forms/.babelrc
@@ -1,17 +1,0 @@
-{
-  "presets": [
-    "@emotion/babel-preset-css-prop",
-    "@babel/preset-typescript",
-    [
-      "@babel/preset-react",
-      {
-        "runtime": "automatic",
-        "importSource": "@emotion/react"
-      }
-    ]
-  ],
-  "plugins": [
-    "babel-plugin-twin",
-    "babel-plugin-macros"
-  ]
-}
