--- packages/forms/.prettierrc.js
@@ -1,14 +1,0 @@
-module.exports = {
-  arrowParens: "avoid",
-  bracketSpacing: true,
-  jsxBracketSameLine: false,
-  printWidth: 80,
-  semi: false,
-  singleQuote: true,
-  tabWidth: 2,
-  trailingComma: "all",
-  useTabs: false,
-  plugins: [
-    require("prettier-plugin-twin.macro")
-  ]
-}
