+++ packages/forms/package.json
@@ -2,6 +2,7 @@
   "name": "forms",
   "version": "1.0.0-pre-alpha.1",
   "main": "index.js",
+  "type": "module",
   "babelMacros": {
     "twin": {
       "preset": "emotion",
@@ -9,51 +10,38 @@
     }
   },
   "scripts": {
-    "dev": "webpack serve --mode=development",
-    "build": "webpack --mode=production",
-    "postbuild": "rm -rf ../../theme/public/dist/forms && mkdir -p ../../theme/public/dist && cp -r public forms && mv forms ../../theme/public/dist/",
-    "serve": "npx serve public"
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "postbuild": "rm -rf ../../theme/public/dist/forms && mkdir -p ../../theme/public/dist && mv ./dist/forms ../../theme/public/dist/",
+    "preview": "vite preview"
   },
   "dependencies": {
     "@emotion/react": "^11.10.5",
     "@emotion/styled": "^11.10.5",
-    "@headlessui/react": "^1.7.5",
+    "@headlessui/react": "^1.7.7",
     "@heroicons/react": "^2.0.13",
     "lodash": "^4.17.21",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-hook-form": "^7.40.0",
-    "react-loading": "^2.0.3"
+    "react-hook-form": "^7.41.1",
+    "react-loading": "^2.0.3",
+    "twin.macro": "^3.1.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.2",
-    "@babel/preset-react": "^7.18.6",
-    "@babel/preset-typescript": "^7.18.6",
-    "@emotion/babel-preset-css-prop": "^11.10.0",
+    "@emotion/babel-plugin": "^11.10.5",
     "@tailwindcss/aspect-ratio": "^0.4.2",
     "@tailwindcss/forms": "^0.5.3",
     "@tailwindcss/line-clamp": "^0.4.2",
     "@tailwindcss/typography": "^0.5.8",
     "@types/lodash": "^4.14.191",
-    "@types/react": "^18.0.25",
+    "@types/node": "^18.11.17",
+    "@types/react": "^18.0.26",
     "@types/react-dom": "^18.0.9",
-    "babel-loader": "^9.1.0",
+    "@vitejs/plugin-react": "^3.0.0",
     "babel-plugin-macros": "^3.1.0",
     "babel-plugin-twin": "^1.1.0",
-    "css-loader": "^6.7.2",
-    "dotenv": "^16.0.3",
-    "downshift": "^7.0.4",
-    "postcss": "^8.4.19",
-    "postcss-preset-env": "^7.8.3",
-    "prettier": "^2.8.0",
-    "prettier-plugin-twin.macro": "^1.0.11",
-    "style-loader": "^3.3.1",
-    "tailwindcss": "^3.2.4",
-    "tsconfig-paths-webpack-plugin": "^4.0.0",
-    "twin.macro": "^3.0.1",
-    "typescript": "4.9.4",
-    "webpack": "^5.75.0",
-    "webpack-cli": "^5.0.0",
-    "webpack-dev-server": "^4.11.1"
+    "typescript": "^4.9.3",
+    "vite": "^4.0.0",
+    "vite-tsconfig-paths": "^4.0.3"
   }
 }
