--- packages/forms/public/index.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-  <title>Twin + Emotion + Webpack + TypeScript Example</title>
-  <meta charset="UTF-8" />
-  <link rel="stylesheet" as="style" crossorigin
-    href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.6/dist/web/static/pretendard-dynamic-subset.css" />
-</head>
-
-<body>
-  <div id="root"></div>
-  <script src="as-inquiry.js"></script>
-</body>
-
-</html>
