+++ packages/forms/src/components/AutoComplete.tsx
@@ -2,22 +2,21 @@
 import { ClassNames } from "@emotion/react";
 import { Transition } from "@headlessui/react";
 import Downshift from "downshift";
-import { Fragment } from "react";
+import { DOMAttributes, Fragment } from "react";
 import tw from "twin.macro";
 import { ChevronDownIcon } from '@heroicons/react/20/solid'
 
-interface SearchBarProps {
+interface SearchBarProps extends DOMAttributes<HTMLDivElement> {
   onChange?: (value: string) => void;
   items: string[];
   initialInputValue?: string;
 }
-const SearchBar = ({ onChange, items, initialInputValue }: SearchBarProps) => {
+const SearchBar = ({ onChange, items, initialInputValue, ...props }: SearchBarProps) => {
   return (
     <Downshift
       onChange={onChange}
       itemToString={item => (item ? item : "")}
       initialInputValue={initialInputValue}
-
     >
       {({
         getInputProps,
@@ -30,14 +29,8 @@ const SearchBar = ({ onChange, items, initialInputValue }: SearchBarProps) => {
         highlightedIndex,
         selectedItem
       }) => (
-        <div>
+        <div {...props}>
           <div tw="m-auto relative h-full">
-            {/* <label
-              {...getLabelProps()}
-              tw="font-bold text-xs text-gray-700 block"
-            >
-              Enter a state {selectedItem}
-            </label> */}
             <input
               tw="relative w-full h-full cursor-default border border-[#6b7280] bg-white py-2 pl-3 pr-10 text-left shadow-sm focus:border-indigo-500 focus:outline-none focus:ring-1 focus:ring-indigo-500 sm:text-sm"
               {...getInputProps()}
@@ -64,7 +57,7 @@ const SearchBar = ({ onChange, items, initialInputValue }: SearchBarProps) => {
                     leaveFrom={css(tw`opacity-100 translate-y-0`)}
                     leaveTo={css(tw`opacity-0 translate-y-1`)}
                   >
-                    <ul tw="absolute z-10 mt-1 max-h-60 w-full overflow-auto rounded-md bg-white py-1 text-base shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none sm:text-sm">
+                    <ul tw="absolute z-10 mt-1 max-h-60 w-full overflow-auto bg-white py-1 text-base shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none sm:text-sm">
                       {items
                         .filter(
                           item =>
@@ -74,7 +67,7 @@ const SearchBar = ({ onChange, items, initialInputValue }: SearchBarProps) => {
                         .map((item, index) => (
                           <li
                             css={[
-                              tw`relative cursor-default select-none py-2 pl-8 pr-4`,
+                              tw`relative cursor-pointer select-none py-2 pl-8 pr-4`,
                               highlightedIndex === index && tw`bg-red-500`
                             ]}
                             {...getItemProps({
