+++ packages/forms/src/form/styleRequire.tsx
@@ -1,0 +1,12 @@
+import { styled } from "twin.macro";
+
+const Require = styled.span`
+color: #ff0000;
+margin-right: 6px;
+line-height: 1;
+display: inline-block;
+margin-top: 0.2em;
+font-size: 20px;
+`
+
+export default Require;
