+++ packages/forms/src/types/vite-env.d.ts
@@ -1,0 +1,2 @@
+/// <reference types="vite/client" />
+/// <reference types="@emotion/react/types/css-prop" />
